NCI-agency/anet#2995: Reuse some of the position showing functionality
