+++ client/src/models/Position.js
@@ -74,6 +74,60 @@ export default class Position extends Model {
 
   static autocompleteQueryWithNotes = `${this.autocompleteQuery} ${GRAPHQL_NOTES_FIELDS}`
 
+  static allFieldsQuery = `
+    uuid
+    name
+    type
+    status
+    code
+    organization {
+      uuid
+      shortName
+      longName
+      identificationCode
+    }
+    person {
+      uuid
+      name
+      rank
+      role
+      avatar(size: 32)
+    }
+    associatedPositions {
+      uuid
+      name
+      type
+      person {
+        uuid
+        name
+        rank
+        role
+        avatar(size: 32)
+      }
+      organization {
+        uuid
+        shortName
+      }
+    }
+    previousPeople {
+      startTime
+      endTime
+      person {
+        uuid
+        name
+        rank
+        role
+        avatar(size: 32)
+      }
+    }
+    location {
+      uuid
+      name
+    }
+    customFields
+    ${GRAPHQL_NOTES_FIELDS}
+  `
+
   static humanNameOfStatus(status) {
     return utils.sentenceCase(status)
   }
